+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v3.1.1 - 2020-07-10
 
 - Reduced health of Tougher Commanders
+- Removed build arm buffs for Tougher Commanders
 
 ## v3.1.0 - 2020-07-09