+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v3.1.1 - 2020-07-10
 
-- Reduced health of Tougher Commanders
+- Reduced health of Tougher Commanders from 1,000,000 to 50,000 (new wars only)
+- Removed build and storage bonuses from Tougher Commanders (new wars only)
 
 ## v3.1.0 - 2020-07-09